Move C++ shim into monad-triedb crate
