+++ rust/crates/monad-triedb/CMakeLists.txt
@@ -27,15 +27,15 @@ endif()
 # again
 add_library(monad_execution_local SHARED IMPORTED)
 set_target_properties(monad_execution_local PROPERTIES
-  INTERFACE_INCLUDE_DIRECTORIES "${PROJECT_SOURCE_DIR}/../../../../"
+  INTERFACE_INCLUDE_DIRECTORIES "${PROJECT_SOURCE_DIR}/../../../"
   IMPORTED_LOCATION "${MONAD_EXECUTION_DIR}/libmonad_execution.so"
 )
 
-add_subdirectory("../../../../" "monad-execution-build"
+add_subdirectory("../../../" "monad-execution-build"
 )# needed for monad-trie
 
-add_library(triedb_driver SHARED "${PROJECT_SOURCE_DIR}/include/triedb.h"
-                                 "${PROJECT_SOURCE_DIR}/src/triedb.cpp")
+add_library(triedb_driver SHARED "${PROJECT_SOURCE_DIR}/include/ffi.h"
+                                 "${PROJECT_SOURCE_DIR}/src/ffi.cpp")
 
 target_include_directories(triedb_driver PUBLIC ${PROJECT_SOURCE_DIR}/include)
