+++ rust/crates/monad-triedb/src/ffi.cpp
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "triedb.h"
+#include "ffi.h"
 
 #include <category/core/byte_string.hpp>
 #include <category/core/log.hpp>
