--- rust/crates/monad-triedb/triedb-driver/.gitignore
@@ -1,1 +1,0 @@
-/build
